+++ app/Services/Contact/Contact/UpdateBirthdayInformation.php
@@ -189,7 +189,7 @@ class UpdateBirthdayInformation extends BaseService
                 'frequency_type' => 'year',
                 'frequency_number' => 1,
                 'title' => trans(
-                    ($data['is_deceased'] ? 
+                    ($data['is_deceased'] ?
                         'people.people_add_birthday_reminder_deceased' : 'people.people_add_birthday_reminder'),
                     ['name' => $contact->first_name]
                 ),