+++ database/migrations/2020_03_22_132429_rename_birthday_reminder_title_deceased.php
@@ -1,8 +1,6 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\Schema;
 
 class RenameBirthdayReminderTitleDeceased extends Migration
 {
@@ -14,11 +12,11 @@ class RenameBirthdayReminderTitleDeceased extends Migration
     public function up()
     {
         $affected = DB::update(
-            "update reminders " .
-            "inner join contacts " .
-                "on reminders.id=birthday_reminder_id " .
-            "set reminders.title = concat('On this date, ', contacts.first_name, ', would have celebrated his birthday') " .
-            "where is_dead = 1;"
+            'update reminders '.
+            'inner join contacts '.
+                'on reminders.id=birthday_reminder_id '.
+            "set reminders.title = concat('On this date, ', contacts.first_name, ', would have celebrated his birthday') ".
+            'where is_dead = 1;'
         );
     }
 }